// src/app/models/rental.model.ts
export class RentalModel {

    id: string;
        nombre: string;
        apellido: string;
        cedula: string;
        correo: string;
        distrito: string;
        ciudad: string;
        carId: string;
        fechaIni?: Date;
        fechaFin?: Date;
        marca: string;
        modelo: string;

        constructor() {}
}